Renamed package to telemetry
